week6_streaming/pyproject.toml
[tool.poetry]
name = "week6-streaming"
version = "0.1.0"
description = ""
authors = ["Carlos Vecina <carlos.vecina@jobandtalent.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
